On lookup matching give less prominence to the search server score.
The search server scores had a high impact on the scoring of lookup results. A bad search server result easily put a match beyond the matching threshold, even if Picard itself considered it a good match. Hence still consider the search server score, but combine it weighted with the rest of the results.