Use sets instead of lists or tuples where applicable
Note: sets aren't always faster, for example if searched element is first in a tuple/list, it can even be very slightly slower.
But sets are always faster when searched element isn't in at all (O(1) vs O(n))
In most cases, it will make no or very little difference, so that's more a stylistic change.